Added mdx function config option
This allows you to change the plugins used depending on the front matter attributes or the file name of the mdx file.

For example, in the docs/*.mdx files you may want to include a table of contents automatically from the headings on the page, but when mdx is imported as a large content section of a different page you don’t want a table of contents.

This commit also adds some solid unit tests around mdx that test the mdx files render as react components.

closes #8