remove build manifest junk from test snapshots
Previously, nearly every change broke several snapshot on nothing more than a bunch of file name hashes from rollup. If we changed the gist app, or changed @remix-run/react, it all changed, which isn’t really the point of the tests.

Now only the content of the page is captured, so the build can do whatever it wants. This isn’t risky because if that stuff doesn’t work, then the other tests will fail (like clicking a link and asserting the new page).

We have specific build tests to test the build output that we can use to aid development, but they aren’t important for CI, so we’ve `describe.skip`ed them.