added `img:` importing
Implementation notes: The whole strategy is based on creating predictable asset names without having to process or even read the image into memory. We want this because every build is actually two builds (server and browser), and during development we don’t want to bog down the build with image processing. We also want the restarts of the dev server (and production build) to use already processed images.