fix vercel deployments
Okay … this was not what I thought it was, hence all the silly commits with various releases before this commit. There was always a problem, but I didn’t realize it and thought a later commit introduced it. Anyway …

First, we were naming layout routes `layout:${route}` to give them a namespace anticipating other routes. I realized that windows wouldn’t like the output of a file with `:` in it so I changed it to `layout__${route}`.

Then I realized that when we sort routes in `react/routes.ts` and construct the route tree, any manual routes that started with any letters before `layout` would sort first and break the whole thing, so I changed it to `__layout__${route}`.

Then I discovered that our Vercel starter couldn’t find the root route in `readConfig` in production. That’s because our build was saving it as `__layout__root.js` and it was looking for `root.js`. On other platforms we deploy both app and build folders, readConfig looks at the original source files to figure out the routes. Since the vercel starter doesn’t deploy the source (only the output) the root route couldn’t be found!

So now we just leave the root route’s id alone (just `root` not `layout:root` or anything else) so that when the file is saved the vercel starter can find it in production.

Then we special case the root route to sort first when constructing the route tree in react/routes.ts so that the later routes can find it as their parent.

Whew!

We really gotta rework this requirement to deploy the source and the build to the server. We have the server-manifest.json, I think `readConfig` should probably just read that in production instead of `appDirectory`.