WIP: keys, not refs, so we can replay and remove code branches
Before we couldn't replay a ref'd submission on browser refresh because refs aren't serializable, they were only in memory. On refresh we didn't know enough to replay.

We also had to cross our fingers with WeakMap that when the app lost a reference to a ref that the garbage collector would get it, since refs don't have a lifecycle.

Because of all this, app developers would have had to write two branches of code if they wanted a solid progressively enhanced experience. This greatly simplifies the code you have to write and keeps you from having to think about "what if the user hits refresh or back from a different domain?".

With serializable keys instead of in-memory refs, developers can't write the code with no branching and Remix knows enough to replay it on refresh/pop-from-different-document.