Add serverTiming option to createRequestHandler
Also add support for the `REMIX_SERVER_TIMING` environment variable
to all node runtimes including:

- `@remix-run/architect`
- `@remix-run/express`
- `@remix-run/vercel`

If `REMIX_SERVER_TIMING` is present it will set the `serverTiming`
option.

BREAKING: this changes the signature of `createRequestHandler` (in
`@remix-run/node`) from

`createRequestHandler(build, mode)`

to

`createRequestHandler({ build, mode, serverTiming })`

But this also brings `@remix-run/node`'s `createRequestHandler` more
in line with the `@remix-run/express` (and architect and vercel)
function of the same name, so I'm ok with it.

BREAKING: this changes the signature of `createApp` from
`@remix-run/serve` from

`createApp(buildDir, mode)`

to

`createApp({ buildDir, mode, serverTiming })`

but again, this signature is more closely aligned with the other public
APIs for creating servers (and probably nobody is using it so far) so
I'm ok with it.