feat: remove `sideEffects: false` requirement
Up to this point we've required all Remix apps to use `{ sideEffects: false }` in their package.json to get esbuild's tree-shaking to remove any loader/action-specific code from the route modules when bundling them for the browser. But with this change we should be able to eliminate this requirement and specify it ourselves for all client app modules.