Fix Variable Used in Conventions.md Code Example
The "Throwing Responses in Loaders" code example appears to use an undefined variable in the CatchBoundary function. The return value of useCatch() is defined as "let caught", but the JSX returned by the boundary tries to use "invoiceCatch" as the variable; "invoiceCatch" does not appear to be defined in the code example. The commit replaces "invoiceCatch" with "caught".