fix: stop sorting params on data fetch
The prefetch link tag and the actual data being fetched were not the same because the data fetching code sorted the params. I'm the dork that sorted it in the first place, it was a mistake.

Now the href's match up and prefetching is actually useful for link with search params:

```jsx
// given this anchor
<Link to="/page?same=thing" />

// prefetch link
<link rel="prefetch" href="/page?same=thing&_data={routeId}" />`

// data fetch
fetch(`/page?same=thing&_data={routeId}`)
```