Bundle code before upload to AWS in arc template
arc-plugin-esbuild bundles and tree-shakes the code before
uploading it to a Lambda function, resulting in much faster Lambda
startups.

Note that the build directory has been moved from `/server/build`
to `/src/build` because arc-plugin-esbuild hardcodes the directory
that it scans for entry points to `/src`. (With the build directory
set to `/server/build`, arc-plugin-esbuild was attempting to
overwrite the file due to this issue.)