Throw error when cookie length exceeds browser max
Session storage inside a cookie is great for smaller things and
normally "just works". For newer developers (and older forgetful ones
like myself), the storage limit can be a source of errors that is
easily missed.

This makes the commit action for cookie session storage throw an
error describing the problem if data is attempted to be persisted that
will exceed the browser max of 4096 bytes.