validate repo for private repos
It looks like GITHUB_TOKEN is ignored for web requests to
private repos: even when that token should grant access, the
server returns a 404.

Instead, we can hit the api, which is happy to accept a token. This
works for public repos as well, with or without a GITHUB_TOKEN. However,
we do need to do a bit of extra work when a filePath is passed. The
api response only tells us some metadata, we need the /tree call to
learn about what files exist there.

A few snapshot error messages changed. I couldn't figure out exactly
why, since both https://github.com/error-username/404 and
https://api.github.com/repos/error-username/404 produce HTTP 404 for me.