Serialize Error name
The name of the error is an important piece of information, which helps
in determining the cause of failures. Usually in Node, the name is also
part of the stack trace, so it may not be seen as completely necessary.
However, some tooling expects the name to be meaningful for diagnostics.
This change serializes the error name in the app state, making it
accessible for tooling which wants to introspect it.