fix(remix-react): prevent form data duplicated value
After : https://github.com/remix-run/remix/commit/5894598c75d0e542651dc92bc400320c108eae8d
A bug was introduced on Safari. Duplicated value of the form button was appended to the form data.

In order to fix this, we check if the value is already in the form data before appending it.
It's not a perfect solution because some use case would probably like to add also the button value even if it's already in the form data.
But since it's a less common use case than just using "Safari" it might be better.