feat(remix-cloudflare-workers): Allow non FetchEvent inputs
When working with Cloudflare ES Module workers[0] we no longer have
access to a `FetchEvent` object. Instead we're given three parameters,
`request`, `environment`, and `context`. This is at odds with the
"Service Worker" API where the environment vars are global and the given
event has a request and helper properties.

To bridge the gap between the two, this change adds a new type `Evt` and
allows that to be passed in lieu of a `FetchEvent`. When this is the
case the user is expected to create the `Evt` object themselves. Here is
an example:

```ts
import { createEventHandler } from "@remix-run/cloudflare-workers";
import type { Evt } from "@remix-run/cloudflare-workers";

import * as build from "../build";

type Environment = Record<string, unknown>;

const handler = createEventHandler({ build });

export default {
 fetch(request: Request, env: Environment, context: ExecutionContext) {
   const event: Evt<Environment> = { request, env, waitUntil: context.waitUntil.bind(context) };
   return handler(event);
 }
}
```

This is very similar what the `@cloudflare/kv-asset-handler` did in
their 0.2.0 release[1];

[0] - https://blog.cloudflare.com/workers-javascript-modules/
[1] - https://github.com/cloudflare/kv-asset-handler/commit/79ebac45c6de6b068a2ac9e8f7fdbce0ade53ba2