refactor(remix-dev): rewrite `jscodeshift`-based migrations as `babel`-based codemods (#4572)
* refactor(dev): extract `useColor` usages into `safe` utility

* refactor(dev): remove outdated reference to compiler shims

* fix(dev): remove access to `convert-to-javascript` migration via CLI

conversion to javascript is not a "migration" in that it does not help the user to upgrade to a
newer version of Remix

* refactor(dev): rewrite `replace-remix-magic-exports` as a babel codemod

Compared to the previous jscodeshift-based migration:
- codemod no longer depends on a network connection
- babel's visitor API for traversing the AST is simpler
- not spinning up workers for applying code transforms

This ends up speeding up the codemod by ~10x and (hopefully 🤞) fixes
some of the issues we were seeing in CI on Windows (since we think
problems are mostly timeouts caused by slow tests or overhead for
workers).

* test(dev): restore fixture

this fixture was incorrectly included as part of dependency upgrades:
- https://github.com/remix-run/remix/pull/3917
- https://github.com/remix-run/remix/pull/3929

it should not have been updated since it is supposed to represent a
Remix 1.3.x codebase

* test(dev): add tests for generic codemods and specifically for `replace-remix-magic-imports`

* ci: run build before primary tests

some tests can make use of the built javascript artifacts
(e.g. `cli.js`) to run 8-10x faster than running with source
Typescript (e.g. `cli.ts`)

* test(dev): retry temp dir removal for windows ci

Windows sometimes throws `EBUSY: resource busy or locked, rmdir`
errors when attempting to removing the temporary directory.
Retrying a couple times seems to get it to succeed.
See https://github.com/jprichardson/node-fs-extra/issues?q=EBUSY%3A+resource+busy+or+locked%2C+rmdir

* Create long-colts-remain.md