fix(server-runtime): ensure create/update is based on underlying data
Currently commitSession doesn't check whether the actual data in the
data storage exists or not. It just does an update if there is an id
stored in the cookie, and a create otherwise.

Note: The data fetch doesn't need to be in a try catch. It was
implemented this way because fileStorage session throws errors when
reading data that isn't there. This isn't handle in this commit to
reduce scope.