bug: v1.8 loaders have empty request.headers when called after action
Prior to Remix v1.8.0 the loader function when called subsequentially to
a (POST) action would have the request's `headers` set. Since v1.8 this
is not the case anymore.

This is highlyt problematic. For example, if you have some high level
session based cookie which is used to guard GET request in the page
action (or a parent layout route); the authentication would fail every
time the loaders data is retrieved subsequentially to a POST action.