Add immutable to Cache-Control header in templates
This is a minor improvement to both the Netlify and Cloudflare Pages templates to include the `immutable` directive in the `Cache-Control` header for the hashed build assets. This removes the `s-maxage` for those as well since they are redundant to the `max-age` given they are equal. These now exactly match `Cache-Control` headers for hashes assets in the Deno, Express, and fly.io (via `remix-serve`) templates.

See `immutable` docs here: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control#immutable