[remix-dev] Specify "paths" in `require.resolve()` check
`require.resolve()` by default will resolve module paths relative to the
`__filename` of the code that is invoking it (so relative to
`serverBareModulesPlugin.js` in this case). This normally works, by
coincidence, because usually `@remix-run/dev` is located within the project's
`node_modules` directory.

However, we get a false positive when `remix` CLI is located *outside*
of the project directory, for example when testing changes in a local
checkout of this repo:

```bash
$ node ~/path/to/remix-run/remix/packages/remix-dev/dist/cli.js build
```

To fix, specify `paths` array to `require.resolve()` and explcitly specify
the `cwd` to resolve from.