fix(remix-dev): require.resolve react-refresh
react-refresh is a dependency of remix-dev, not of your app.

so we can use require.resolve to find it. This fixes an issue where remix-dev would not work with pnpm.

see https://github.com/pnpm/pnpm/issues/2957#issuecomment-719832332 for a similar issue in CRAv4 and how they fixed it there https://github.com/pmmmwh/react-refresh-webpack-plugin/pull/230/files

Signed-off-by: Logan McAnsh <logan@mcan.sh>