refactor(dev): revert serving assets from dev server
When sending HMR updates, the assumption was that serving assets from
dev server would allow those updates to be sent immediately, without
waiting for the app server to restart. But this isn't the case (e.g. if
you navigate to another page that has a loader and the app server is
rebooting, that loader fetch will fail).