docs: rename "dev server" to "Remix compiler"
While `remix dev` technically spins up a "dev server", it is not a dev
server in the traditional sense; it does not replace your app server in
dev. Rather the "dev server" is just the Remix compiler running in watch
mode but with a single HTTP endpoint (/ping) for coordinating hot
updates.