Make remix-express agnostic of express
Remix depends on express without using most of its features
and mostly relying on node http server inside of it.

Express is quite big package which may not be necessary in most apps.
https://packagephobia.com/result?p=express

Here I replaced express types and covered missing features.
Later we can do the same with serve and dev.