fix(remix-dev): fix server builds where serverBuildPath ends in .cjs
Fix a bug that caused the server build file to be emitted into the
assets directory if the value of `serverBuildPath` ended in `.cjs`.

The bug was due to the server-side portion of the Remix compiler
only iterating over files ending in `.js` and `.mjs`, but not
`.cjs`.

Note that this diff is smaller than it appears: the largest number
of lines affected are in integration/compiler-mjs-cjs-output-test.ts,
in which I increased the indentation of the test fixture code in
order to parametrize it over file extension.