fix: apply workaround for 'dynamic require not supported' error
Workaround given here is to define 'require' using the 'banner' esbuild
option when building for esm and node.

https://github.com/evanw/esbuild/issues/1921#issuecomment-1439609735

Used the async IIFE form to avoid import collisions, as the comment says.
Skipped __filename and __dirname, as I found existing shims for those two
in the codebase which caused collisions.

The timeline and discussion on that issue shows a similar workaround being
applied in several other projects.