vite: no `enforce: "post"`
"Post" plugins are meant primarily for build analysis.
Previously we were relying on `enforce: "post"` to ensure that plugins
that transpile other languages to JS/TS ran before the Remix plugin.
However, it should be the responsability of _those_ plugins to declare
themselves with `enforce: "pre"`. In the case of MDX, the canonical
plugin is not a Vite plugin, but rather a Rollup plugin so it cannot set
`enforce` as that is a Vite-only API.

Ideally, someone would wrap the MDX Rollup plugin and declare it as
`enforce: "pre"`. This can also be done in userland through a simple
object spread.

That said, its simpler to place the MDX plugin before the Remix plugin
in the Vite plugins array. Order is semantically meaningful for Vite
plugins anyway.