fix: move manifestPath to use vite's assetsDir
Previously the manifest location was set to the hardcoded path of `assets`, but if a user configures their build to change the `assetsDir` from Vite the result was a file that was still left in the assets folder.

To handle the configuration in Vite, this moves the path generation to utilize Vite's value so that the manifest javascript file is colocated with all other client assets.

Signed-off-by: Eric Chernuka <ericchernuka@gmail.com>