wip
I think `vite-basename-test` tests that are failing are sharing
resources that should be isolated. Right now we're relying on
side-effects for setting `port` and `cwd` between tests which I think is
the root cause of the error especially since the only test for `express
build` that passes is the one that _doesn't_ share resources with other
tests.

Haven't looked into other failures, but my guess is its the same thing.
Probably would be fixed if we use our new Vite test fixtures.