Ignore only local env files in Vite templates
In Vite `.env` files work a bit differently, where `.env` itself is
intended to be public, and `.env.local` is the one that should be
ignored by Git. But a Vite project can have more `.env` files, like
`.env.development`, which is used only in development mode, and
overrides values in `.env`, then you can have `.env.development.local`
as well and so on.

https://vitejs.dev/guide/env-and-mode.html#env-files