Remove unnecessary useEffect from tutorial code example
Using useEffect for syncing state with props (or other external state) is unnecessary, and can trigger additional renders and potentially flashes of incorrect content. We can use a useState to detect changes in order to trigger a sync. 

This is explicitly allowed by React – https://react.dev/learn/you-might-not-need-an-effect#adjusting-some-state-when-a-prop-changes