Docs: Clarify `useActionData` Hook Scope
The current documentation for the `useActionData` hook might
unintentionally imply that it can be used to access data from any route
action. However, the hook is limited to accessing data associated with
the current route action where it's called.

This PR clarifies this limitation in the documentation to prevent
confusion.

See #9717 for more information.