Add property initializers to SuperHeaders
This allows SuperHeaders to be initialized with an object of property/value pairs that correspond with the names of the getter/setter methods for each header.

Also, added a bunch of doc comments to various classes and methods and ensured that all classes can be initialized with an instance of themselves.