fix: don't Base64 encode cookie if encode is set
Right now, if a developer adds an encode/decode function when creating a cookie, it's applied after the cookie is Base64 encoded. We should not Base64 encode a cookie if the developer have added a `encode`/`decode` override to the
cookie options. Unfortunately this makes the override and sign/secret options mutually exclusive.

BREAKING CHANGE: If you override the cookie encode/decode functions, you can't sign the cookie with a secret