Iterate manually over `response.body`
I'm honestly not sure if there is a problem in Node itself or in undici, but I'm seeing issues with the stream reader not releasing its lock properly when using `for await...of` to iterate over `response.body` in `sendResponse`. In any case, there may be a variety of `Response` implementations that people eventually want to use with this, and async iterators are hard to get right, so dropping down to the lower-level `getReader()` API for stream iteration seems like a prudent move.