route-pattern: fix type distribution for `Parse`
At first, `Parse` used `T extends any ? ... : ...` to force TS to
perform union distribution. Later, that was changed to `[T] extends [string] ? ... : ...`
as an "optimization", but in reality that change is a no-op since we already check for `<T
extends string>` in the generic declaration.
As a result, `Parse` no longer performed union distribution which caused
it to incorrectly compute types when unions were passed in.
Additionally, the benchmarks for `Parse` were incorrect since they only
reflected the type instantiations when incorrectly computing parsed
types.

To fix this, a test was added that explicitly passes in a union type to
`Parse` and the corresponding type benchmarks were updated. The new test
unsurprisingly fails with `[T] extends [string] ? ...: ...` and passed
with `T extends any ? ... : ...`.